Save states 'n shizz.

+++ src/fceumm/ppu.c
@@ -63,8 +63,8 @@ int FCEU_InitVirtualVideo(void)
 
 	if(sizeof(uint8*)==4)
 	{
-		uint32 m;
-		m=(uint32)XBuf;
+		uintptr_t m;
+		m=(uintptr_t)XBuf;
 		m=(4-m)&3;
 		XBuf+=m;
 	}
